slack_notion_integration/requirements.txt
slack-sdk==3.21.3
slack-bolt==1.18.0
notion-client==2.0.0
pytz==2023.3
python-dotenv==0.20.0
flask
